requirements.txt
google-generativeai>=0.3.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
flask>=3.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
openpyxl>=3.1.0

